Limits website middleware path matching of periods
The current implementation doesn't allow for path matching on routes such as: `/accounts/register/:token`, where token is a JWT that has a standard period in it.  

By changing the matcher to check for a period (dot) only inside of the path, this allows us to pass params containing periods in the query.